+++ Source/Buttons/readyconsumable.js
@@ -24,6 +24,7 @@ module.exports = new Button(id, (interaction, args) => {
 						new MessageSelectMenu().setCustomId(`consumable${SAFE_DELIMITER}${adventure.room.round}`)
 							.setPlaceholder("Pick a consumable...")
 							.addOptions(Object.keys(adventure.consumables).reduce((options, consumable) => options.concat({
+								//TODONOW handle more than 25 types of consumables in inventory
 								label: `${consumable} (Held: ${adventure.consumables[consumable]})`,
 								description: getConsumable(consumable).description,
 								value: consumable