+++ Source/consumables/explosionpotion.js
@@ -4,7 +4,7 @@ const { dealDamage } = require("../combatantDAO.js");
 module.exports = new ConsumableTemplate("Explosion Potion", "Deal 75 damage to all enemies", selectTargets, effect)
 	.setElement("Untyped")
 	.setTargetTags("all", "enemy")
-	.setFlavorText([]);
+	.setFlavorText(["*Additional Notes*", "*Not to be confused with __Fiery Potion__. DO NOT apply to self.*"]);
 
 function selectTargets(userIndex, adventure) {
 	// all enemies