+++ Source/equipment/potion-base.js
@@ -1,6 +1,7 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { addModifier } = require('../combatantDAO.js');
 
+//TODONOW convert to potion kit
 module.exports = new EquipmentTemplate("Unfinished Potion", 1, "*Apply @{mod1Stacks} @{mod1} to a foe*\nCritical Hit: @{mod1} x@{critBonus}", "Water", effect, ["Earthen Potion", "Windy Potion", "Watery Potion"])
 	.setCategory("Trinket")
 	.setTargetingTags({ target: "single", team: "enemy" })